backend/pyproject.toml
[project]
name = "assitance"
version = "0.1.0"
description = "Multi-purpose AI assistant with flexible model support"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2",
    "alembic>=1.14",
    "pydantic-settings>=2.0",
    "openai>=1.40",
    "anthropic>=0.34",
    "ollama>=0.4",
    "python-multipart>=0.0.9",
    "cryptography>=44.0",
    "duckduckgo-search>=6.0",
    "websockets>=13.0",
    "google-genai>=1.65.0",
    "chromadb>=1.5.2",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.2.3",
    "playwright>=1.58.0",
    "markdownify>=1.2.2",
    "redis[hiredis]>=5.0",
    "httpx>=0.27",
    "tiktoken>=0.12.0",
    "edge-tts>=7.2.7",
    "docker>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
